fix: pin python-daemon<3.0.0 to avoid build errors
Builds of projects that require `ansible-runner` and thereby also
require `python-daemon>=3.0.0` in environments with modern versions
of `setuptools` are failing with this error:

```
ModuleNotFoundError: No module named 'setuptools.extern'
```

The problematic `extern` import in `python-daemon` was introduced in
its `3.0.0` release and has since been fixed in that project's `main`
branch, but there is not yet any release including that fix. Simply
pinning an older version appears to resolve this build issue and does
not appear to introduce any regressions.

After `python-daemon` releases a new version, assuming they do, then
we probably should change this from `<3.0.0` to `>3.0.1`.

See also:

* https://pagure.io/python-daemon/issue/92
* https://github.com/quipucords/quipucords/pull/2714